restructured my repo
