+++ завдання/Лабораторна робота 1.md
