+++ завдання/Лабораторна робота 2.md
