+++ завдання/Лабораторна робота 5.md
