+++ завдання/Лабораторна робота 6.md
